+++ src/Interval.hpp
@@ -4,15 +4,15 @@
 
 class Interval {
 public:
-    double min, max;
+    float min, max;
 
     Interval():
-        min(+std::numeric_limits<double>::infinity()),
-        max(-std::numeric_limits<double>::infinity())
+        min(+std::numeric_limits<float>::infinity()),
+        max(-std::numeric_limits<float>::infinity())
     {
     } // Default interval is empty
 
-    Interval(double min_, double max_):
+    Interval(float min_, float max_):
         min(min_),
         max(max_)
     {
@@ -25,13 +25,13 @@ public:
         max = a.max >= b.max ? a.max : b.max;
     }
 
-    double size() const { return max - min; }
+    float size() const { return max - min; }
 
-    bool contains(double x) const { return min <= x && x <= max; }
+    bool contains(float x) const { return min <= x && x <= max; }
 
-    bool surrounds(double x) const { return min < x && x < max; }
+    bool surrounds(float x) const { return min < x && x < max; }
 
-    double clamp(double x) const
+    float clamp(float x) const
     {
         if (x < min)
             return min;
@@ -40,7 +40,7 @@ public:
         return x;
     }
 
-    Interval expand(double delta) const
+    Interval expand(float delta) const
     {
         auto padding = delta / 2;
         return Interval(min - padding, max + padding);
@@ -50,13 +50,13 @@ public:
 };
 
 const Interval Interval::empty =
-    Interval(+std::numeric_limits<double>::infinity(), -std::numeric_limits<double>::infinity());
+    Interval(+std::numeric_limits<float>::infinity(), -std::numeric_limits<float>::infinity());
 const Interval Interval::universe =
-    Interval(-std::numeric_limits<double>::infinity(), +std::numeric_limits<double>::infinity());
+    Interval(-std::numeric_limits<float>::infinity(), +std::numeric_limits<float>::infinity());
 
-Interval operator+(const Interval &ival, double displacement)
+Interval operator+(const Interval &ival, float displacement)
 {
     return Interval(ival.min + displacement, ival.max + displacement);
 }
 
-Interval operator+(double displacement, const Interval &ival) { return ival + displacement; }
+Interval operator+(float displacement, const Interval &ival) { return ival + displacement; }
